requirements.txt
fastmcp>=0.1.0
pillow>=10.0.0
pyscreenshot>=3.1